+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `ActiveModel::Attributes.attribute_names` now takes into account attributes declared in `.delegate_attribute`
 - relax rubocop config.
+- refactor: extract persistence module
 
 ## [0.11.1] - 2025-06-10